move: 디렉토리 이동 및 그에 따른 변경사항
