+++ src/auth-service/src/main/java/com/phishing/authservice/payload/request/EditProfileRequest.java
@@ -1,4 +1,4 @@
-package com.phishing.authservice.dto.request;
+package com.phishing.authservice.payload.request;
 
 public record EditProfileRequest(
         String nickname,
