+++ src/auth-service/src/main/java/com/phishing/authservice/service/UserService.java
@@ -2,8 +2,8 @@ package com.phishing.authservice.service;
 
 import com.phishing.authservice.component.token.TokenResolver;
 import com.phishing.authservice.domain.User;
-import com.phishing.authservice.dto.request.EditProfileRequest;
-import com.phishing.authservice.dto.request.SignUpRequest;
+import com.phishing.authservice.payload.request.EditProfileRequest;
+import com.phishing.authservice.payload.request.SignUpRequest;
 import com.phishing.authservice.exception.exceptions.DuplicateEmailException;
 import com.phishing.authservice.payload.MemberInfo;
 import com.phishing.authservice.repository.UserRepository;
@@ -54,12 +54,10 @@ public class UserService {
                 request.nickname(),
                 request.phnum()
         );
-        System.out.println("targetUser = " + targetUser.getNickname());
     }
 
     private MemberInfo getMemberInfoToToken(HttpServletRequest request) {
         String token = request.getHeader("Authorization");
-        MemberInfo result = tokenResolver.getClaims(token);
-        return result;
+        return tokenResolver.getClaims(token);
     }
 }
